+++ server/app/services/__init__.py
@@ -7,9 +7,11 @@
 from app.services.auth_service import AuthService
 from app.services.player_service import PlayerService
 from app.services.weapon_service import WeaponService
+from app.services.monster_service import MonsterService
 
 __all__ = [
     "AuthService",
     "PlayerService",
-    "WeaponService"
+    "WeaponService",
+    "MonsterService"
 ]